\name{new.fetcher}
\alias{new.fetcher}
\title{Create a new easy handle with data to allow it to signal completion
and extract content}
\usage{
  new.fetcher(url)
}
\arguments{
  \item{url}{the url to download}
}
\value{
  a new fetcher
}
\description{
  Create a new easy handle with data to allow it to signal
  completion and extract content
}

