\name{queue_deferred}
\alias{queue_deferred}
\title{Tell the multi controller to start a new url for download}
\usage{
  queue_deferred(deferred_httr, url, murl)
}
\arguments{
  \item{deferred_httr}{the deferred httr request}

  \item{url}{that is requested or a descriptive name}

  \item{murl}{the multi url controller}
}
\value{
  a wrapper around a deferred http handle.

  more documentation required.
}
\description{
  Tell the multi controller to start a new url for download
}

