\name{queue_download}
\alias{queue_download}
\title{Tell the multi controller to start a new url for download}
\usage{
  queue_download(url, murl)
}
\arguments{
  \item{url}{the url to download}

  \item{murl}{the multi url controller}
}
\value{
  a wrapper around a deferred http handle.

  more documentation required.
}
\description{
  Tell the multi controller to start a new url for download
}

